refactor(dispatcher): remove CoreEventEmitter.js
