+++ src/index.js
@@ -5,11 +5,9 @@ import Store from "./Store";
 import StoreGroup from "./UILayer/StoreGroup";
 import UseCase from "./UseCase";
 import Context from "./Context";
-import CoreEventEmitter from "./CoreEventEmitter";
 // re-export
 module.exports.Dispatcher = Dispatcher;
 module.exports.Store = Store;
 module.exports.StoreGroup = StoreGroup;
 module.exports.UseCase = UseCase;
 module.exports.Context = Context;
-module.exports.CoreEventEmitter = CoreEventEmitter;
